First pass jumpstart.md
Is there anything I can do to make content for the sections that say -- This feature is not yet supported or not documented?
I'm not familiar with the :::tab notation so I have to look that up if it's part of standard md. Or maybe it's an extension? Guessing it just puts a content under a tab yeah?
Anyway just did a quick edit to clean it up.